refactor(web): group cms components (part 1) - #95
