+++ apps/web/src/lib/components/cms/Hero/Hero.svelte
