+++ apps/web/src/lib/components/cms/Hero/index.ts
