+++ apps/web/src/lib/components/cms/Hero/types.ts
