--- apps/web/src/lib/components/index.ts
@@ -1,13 +1,0 @@
-export * from './Alert';
-export * from './Content';
-export * from './EmbeddedEvents';
-export * from './Footer';
-export * from './Header';
-export * from './Hero';
-export * from './Img';
-export * from './Link';
-export * from './LoadingOverlay';
-export * from './Media';
-export * from './Menu';
-export * from './SectionTitle';
-export * from './Video';
