--- apps/web/src/lib/stores/index.ts
@@ -1,2 +1,0 @@
-export * from './dom';
-export * from './loading';
