--- apps/web/src/lib/types/index.ts
@@ -1,1 +1,0 @@
-export * from './data-list';
