--- apps/web/src/lib/utils/index.ts
@@ -1,7 +1,0 @@
-export * from './calculate-duration';
-export * from './local-storage';
-export * from './resolve-events';
-export * from './resolve-hero-variant';
-export * from './resolve-link-appearance';
-export * from './resolve-link-href';
-export * from './resolve-resource-size';
